Add env OLC_DUMP_SAVE_TMP_DIR to support saving of temperary dir

+++ driver/olCompiling/hip_utility/tmp_dir.cpp
@@ -29,6 +29,8 @@
 #include <boost/filesystem.hpp>
 #include <logger.hpp>
 
+OLC_DECLARE_ENV_VAR(OLC_DEBUG_SAVE_TEMP_DIR)
+
 namespace olCompile {
 
 void SystemCmd(std::string cmd)
@@ -55,7 +57,10 @@ void TmpDir::Execute(std::string exe, std::string args) const
 
 TmpDir::~TmpDir()
 {
-    boost::filesystem::remove_all(this->path);
+    if( !olCompile::IsEnabled(OLC_DEBUG_SAVE_TEMP_DIR{}) )
+    {
+       boost::filesystem::remove_all(this->path);
+    };
 }
 
 } // namespace olCompile
